feat: expose deployment lifecycle sub_step on v2 GQL/REST surfaces
The v1 ``DeploymentDTO.sub_step`` had no v2 counterpart, so the GQL
``ModelDeployment`` node and the v2 REST ``DeploymentNode`` were missing
the rolling-update progress signal. Promote ``DeploymentLifecycleSubStep``
to ``common.data.model_deployment.types`` so the v2 DTO layer can reach it
(``common/`` cannot import from ``manager/``), repoint every existing
importer at the common location, then surface the enum on:

- ``DeploymentNode.sub_step`` (v2 REST DTO)
- ``ModelDeployment.subStep`` (GQL, backed by a new
  ``DeploymentLifecycleSubStep`` GQL enum)

The v2 adapter pipes ``ModelDeploymentData.sub_step`` through to the
new DTO field, and the v2 schema dump picks up the enum + field.
