+++ docs/manager/graphql-reference/supergraph.graphql
@@ -5234,6 +5234,17 @@ enum DeploymentHistoryOrderField
   UPDATED_AT @join__enumValue(graph: STRAWBERRY)
 }
 
+"""
+Added in UNRELEASED. Sub-step within a deployment's current lifecycle phase. ``None`` when the deployment is not in a sub-step-bearing phase.
+"""
+enum DeploymentLifecycleSubStep
+  @join__type(graph: STRAWBERRY)
+{
+  DEPLOYING_PROVISIONING @join__enumValue(graph: STRAWBERRY)
+  DEPLOYING_ROLLING_BACK @join__enumValue(graph: STRAWBERRY)
+  DEPLOYING_COMPLETED @join__enumValue(graph: STRAWBERRY)
+}
+
 """Added in UNRELEASED. Deployment options payload response."""
 type DeploymentOptionsInfoGQL
   @join__type(graph: STRAWBERRY)
@@ -9332,6 +9343,11 @@ type ModelDeployment implements Node
   Added in UNRELEASED. Replica scaling axis, orthogonal to ``metadata.status`` (lifecycle). ``SCALING`` while the replica reconciler is adjusting replica count; ``STABLE`` once holding at the desired count.
   """
   scalingState: ScalingState!
+
+  """
+  Added in UNRELEASED. Sub-step within the current lifecycle phase. ``None`` when the deployment is not in a sub-step-bearing phase.
+  """
+  subStep: DeploymentLifecycleSubStep
 
   """
   Added in 26.4.3. The current active revision of this deployment, resolved via DataLoader.
