+++ src/ai/backend/manager/sokovan/deployment/strategy/applier.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 import logging
 from dataclasses import dataclass, field
 
+from ai.backend.common.data.model_deployment.types import DeploymentLifecycleSubStep
 from ai.backend.common.identifier.deployment import DeploymentID
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.data.deployment.types import (
-    DeploymentLifecycleSubStep,
     RouteStatus,
     RouteTrafficStatus,
 )
